requirements.txt
beautifulsoup4
fpdf2
python-jose
requests
eel
Pillow
